more work on type inference
I am still grappling a bit with hindley-milner and the addition
of the qualifiers. For example, where should the qualifiers be
enforced - during the generation of constraints or when unifying
them?
Also, how pattern matching interacts with HM is not well described
in Stephens tutorial, so this is also a question (but one with an
answer).